fix 0 test

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -10,7 +10,7 @@ unsigned int binary_to_uint(const char *b)
 {
 	int len;
 	int i;
-	unsigned long int array[1024];
+        int array[1024];
 	unsigned long int sum = 0;
 
 	if (b == NULL)
